Move describe logic to pkg/component/describe
